backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
supabase==2.15.2
google-genai==1.10.0
python-dotenv==1.1.0
pydantic==2.11.2
